quick change from code review comments

+++ src/SFA.DAS.EAS.Employer_Financial.Database/StoredProcedures/GetAccountBalance_ByAccountIds.sql
@@ -1,6 +1,6 @@
 ﻿CREATE PROCEDURE [employer_financial].[GetAccountBalance_ByAccountIds]
 	@AccountIds [employer_financial].[AccountIds] Readonly,
-	@allowancePercentage FLOAT = 0
+	@allowancePercentage DECIMAL = 0
 AS
 	SELECT
 		acc.AccountId,
